% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoShiny.R
\name{demoShiny}
\alias{demoShiny}
\title{demoShiny}
\usage{
demoShiny(topic)
}
\arguments{
\item{topic}{character: the topic which should be demonstrated}
}
\value{
nothing
}
\description{
Runs a Shiny app as demo or shows an overview of all Shiny demo apps related to a \code{topic}. 
If more than one app relates to \code{topic} a list of apps is returned otherwise the app is run.
For more details see \code{vignettes('demoShiny')}.
}
\examples{
# collect all apps of loaded packages
demoShiny() 
# collect all apps of the package demoShiny or with the name demoShiny
demoShiny('demoShiny') 
\dontrun{
  demoShiny('demoShiny::app1') 
  demoShiny('silhouette') 
  # use partial matching
  demoShiny('silh') 
}
}
